trans-excel-requirements.txt
openai>=1.0.0
xlwings>=0.30.0
python-dotenv>=1.0.0
pathlib>=1.0.1
streamlit>=1.32.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl-image-loader
openpyxl